increase and decrease cart items done

+++ include/decrease.php
@@ -3,13 +3,16 @@
     $pid = $_GET['pid'];
     $ip = getIp();
     $qty = $_GET['quan'];
-    $q = $qty + 1;
-    /*$delete_item = "delete from cart where pid = '$pid' and ipadd = '$ip'";
-    $run_delete = mysqli_query($con, $delete_item);
-    $insert_pro = "insert into cart(pid,ipadd,qty) values('$pid','$ip','$q')";
-    $run_pro = mysqli_query($con, $insert_pro);*/
-	$update_try = "update cart set qty = '$q' where pid = '$pid' and ipadd = '$ip'";
+    $q = $qty - 1;
+    if($qty>1){
+        $update_try = "update cart set qty = '$q' where pid = '$pid' and ipadd = '$ip'";
     $run_update = mysqli_query($con, $update_try);
+    }
+    if($qty==1){
+        $delete_item = "delete from cart where pid = '$pid' and ipadd = '$ip'";
+        $run_delete = mysqli_query($con, $delete_item);
+    }
+
